♻️ Refactor thumbnail jobs
This commit will overhaul the thumbnail jobs to remove duplication.  We
will leverage inheritance to reduce the amount of code needed.  The only
main difference between the pdf and image job is that in the pdf job, we
need to convert the first page of the pdf into an image and from there
generate a thumbnail.